Trying to fix briefing notes

+++ iira6001/make.sh
@@ -15,4 +15,7 @@ jupyter-book build .
 
 
 #rsync -av --delete _build/html/ hasc@login.ansatt.ntnu.no:/home/groupswww/iirevu/iira6001/
-rsync -rv --delete _build/html/ login.ansatt.ntnu.no:/home/groupswww/iirevu/iira6001/
+if test x$1 = x-i
+then
+   rsync -rv --delete _build/html/ login.ansatt.ntnu.no:/home/groupswww/iirevu/iira6001/
+fi
